+++ src/Atoms/CssGrid/utils.test.ts
@@ -1,4 +1,4 @@
-import { getAreasInfo } from './utils';
+import { getAreasInfo, getMsRawTemplateColumnOrRowStyles } from './utils';
 
 describe('getAreasInfo', () => {
   test('Areas #1 w/gap ', () => {
@@ -84,5 +84,29 @@ describe('getAreasInfo', () => {
       sidebar: { id: 5, colStart: 5, colSpan: 1, rowStart: 5, rowSpan: 5 },
       content: { id: 6, colStart: 3, colSpan: 1, rowStart: 9, rowSpan: 1 },
     });
+  });
+});
+
+describe('getMsRawTemplateColumnOrRowStyles', () => {
+  test('Add gutter between spaces', () => {
+    const raw = '1fr 1fr 1fr';
+    const gutter = '20px';
+
+    const result = getMsRawTemplateColumnOrRowStyles(raw, gutter);
+    expect(result).toEqual('1fr 20px 1fr 20px 1fr');
+  });
+  test('dont touch spaces inside parentheses', () => {
+    const raw = '1fr minmax(100px, 3fr) 1fr';
+    const gutter = '20px';
+
+    const result = getMsRawTemplateColumnOrRowStyles(raw, gutter);
+    expect(result).toEqual('1fr 20px minmax(100px, 3fr) 20px 1fr');
+  });
+  test('dont add extra gutters if multiple spaces', () => {
+    const raw = '10ch        412px    50%  1fr';
+    const gutter = '30px';
+
+    const result = getMsRawTemplateColumnOrRowStyles(raw, gutter);
+    expect(result).toEqual('10ch 30px 412px 30px 50% 30px 1fr');
   });
 });